Moved data splitter in dh_segment_torch
