+++ dh_segment_torch/data/__init__.py
@@ -1,4 +1,5 @@
 from dh_segment_torch.data.color_labels import *
 from dh_segment_torch.data.data_loader import *
+from dh_segment_torch.data.data_splitter import *
 
-__all__ = ["ColorLabels", "DataLoader"]
+__all__ = ["ColorLabels", "DataLoader", "DataSplitter"]
